requirements.txt
asgiref==3.8.1
Django==5.0.4
django-mdeditor==0.1.20
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1
